Optimized the images; using data-uri for nav images.
When I checked out the site, I noticed the background image seemed a
little heavy for the content and the nav images would flicker initially
as the CSS was loading the hover images.

Between using a sprite and data-uri for the nav images, the latter
seemed more straight forward and uses fewer http requests. I could see a
case for either tho.